+++ src/utils.cpp
@@ -70,7 +70,7 @@ std::string module_from_path(const std::string& path) {
 
 std::string mangle_path(std::string path) {
     std::string out;
-    for (int i = 0; i < path.size(); i++) {
+    for (size_t i = 0; i < path.size(); i++) {
         if (path[i] == '.') {
             out += "_D_";
         } else {
